Split services and stateful sets into separate files
