+++ internal/controller/memgraphha_controller.go
@@ -45,31 +45,13 @@ The ResourceList type will have to be used to define the limits and requests of 
 
 import (
 	"context"
-	"fmt"
 
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-
-	"github.com/go-logr/logr"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	// this corresponds to cachev1alpha1 from Memcached example
 	memgraphv1 "github.com/memgraph/kubernetes-operator/api/v1"
 )
-
-// MemgraphHAReconciler reconciles a MemgraphHA object
-type MemgraphHAReconciler struct {
-	client.Client
-	Scheme *runtime.Scheme
-}
 
 //+kubebuilder:rbac:groups=memgraph.com,resources=memgraphhas,verbs=get;list;watch;create;update;patch;delete
 //+kubebuilder:rbac:groups=memgraph.com,resources=memgraphhas/status,verbs=get;update;patch
@@ -177,662 +159,6 @@ func (r *MemgraphHAReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 
 	// The resource doesn't need to be reconciled anymore
 	return ctrl.Result{}, nil
-}
-
-func (r *MemgraphHAReconciler) reconcileDataInstanceNodePortService(ctx context.Context, memgraphha *memgraphv1.MemgraphHA, logger *logr.Logger, dataInstanceId int) (bool, error) {
-	serviceName := fmt.Sprintf("memgraph-data-%d-external", dataInstanceId)
-	logger.Info("Started reconciling NodePort service", serviceName)
-
-	dataInstanceNodePortService := &corev1.Service{}
-	err := r.Get(ctx, types.NamespacedName{Name: serviceName, Namespace: memgraphha.Namespace}, dataInstanceNodePortService)
-
-	if err == nil {
-		logger.Info("NodePort", serviceName, "already exists.")
-		return false, nil
-	}
-
-	if errors.IsNotFound(err) {
-		nodePort := r.createDataInstanceNodePort(memgraphha, dataInstanceId)
-		logger.Info("Creating a new NodePort", "NodePort.Namespace", nodePort.Namespace, "NodePort.Name", nodePort.Name)
-		err := r.Create(ctx, nodePort)
-		if err != nil {
-			logger.Error(err, "Failed to create new NodePort", "NodePort.Namespace", nodePort.Namespace, "NodePort.Name", nodePort.Name)
-			return true, err
-		}
-		logger.Info("NodePort", serviceName, "is created.")
-		return true, nil
-	}
-
-	logger.Error(err, "Failed to fetch NodePort", serviceName)
-	return true, err
-
-}
-
-func (r *MemgraphHAReconciler) createDataInstanceNodePort(memgraphha *memgraphv1.MemgraphHA, dataInstanceId int) *corev1.Service {
-	serviceName := fmt.Sprintf("memgraph-data-%d-external", dataInstanceId)
-	dataInstanceName := fmt.Sprintf("memgraph-data-%d", dataInstanceId)
-	// TODO: (andi) Extract somehow configuration and move into separate files.
-	boltPort := 7687
-
-	dataInstanceNodePort := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serviceName,
-			Namespace: memgraphha.Namespace,
-		},
-		Spec: corev1.ServiceSpec{
-			Type:     corev1.ServiceTypeNodePort,
-			Selector: createDataInstanceLabels(dataInstanceName),
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "bolt",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(boltPort),
-					TargetPort: intstr.FromInt(boltPort),
-				},
-			},
-		},
-	}
-
-	ctrl.SetControllerReference(memgraphha, dataInstanceNodePort, r.Scheme)
-	return dataInstanceNodePort
-}
-
-func (r *MemgraphHAReconciler) reconcileDataInstanceClusterIPService(ctx context.Context, memgraphha *memgraphv1.MemgraphHA, logger *logr.Logger, dataInstanceId int) (bool, error) {
-	serviceName := fmt.Sprintf("memgraph-data-%d", dataInstanceId)
-	logger.Info("Started reconciling ClusterIP service", serviceName)
-
-	dataInstanceClusterIPService := &corev1.Service{}
-	err := r.Get(ctx, types.NamespacedName{Name: serviceName, Namespace: memgraphha.Namespace}, dataInstanceClusterIPService)
-
-	if err == nil {
-		logger.Info("ClusterIP", serviceName, "already exists.")
-		return false, nil
-	}
-
-	if errors.IsNotFound(err) {
-		clusterIP := r.createDataInstanceClusterIP(memgraphha, dataInstanceId)
-		logger.Info("Creating a new ClusterIP", "ClusterIP.Namespace", clusterIP.Namespace, "ClusterIP.Name", clusterIP.Name)
-		err := r.Create(ctx, clusterIP)
-		if err != nil {
-			logger.Error(err, "Failed to create new ClusterIP", "ClusterIP.Namespace", clusterIP.Namespace, "ClusterIP.Name", clusterIP.Name)
-			return true, err
-		}
-		logger.Info("ClusterIP", serviceName, "is created.")
-		return true, nil
-	}
-
-	logger.Error(err, "Failed to fetch ClusterIP", serviceName)
-	return true, err
-
-}
-
-func (r *MemgraphHAReconciler) createDataInstanceClusterIP(memgraphha *memgraphv1.MemgraphHA, dataInstanceId int) *corev1.Service {
-	serviceName := fmt.Sprintf("memgraph-data-%d", dataInstanceId)
-	dataInstanceName := serviceName
-	// TODO: (andi) Extract somehow configuration and move into separate files.
-	boltPort := 7687
-	replicationPort := 20000
-	mgmtPort := 10000
-
-	dataInstanceClusterIP := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serviceName,
-			Namespace: memgraphha.Namespace,
-		},
-		Spec: corev1.ServiceSpec{
-			Type:     corev1.ServiceTypeClusterIP,
-			Selector: createDataInstanceLabels(dataInstanceName),
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "bolt",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(boltPort),
-					TargetPort: intstr.FromInt(boltPort),
-				},
-				{
-					Name:       "replication",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(replicationPort),
-					TargetPort: intstr.FromInt(replicationPort),
-				},
-				{
-					Name:       "management",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(mgmtPort),
-					TargetPort: intstr.FromInt(mgmtPort),
-				},
-			},
-		},
-	}
-
-	ctrl.SetControllerReference(memgraphha, dataInstanceClusterIP, r.Scheme)
-	return dataInstanceClusterIP
-}
-
-func (r *MemgraphHAReconciler) reconcileDataInstance(ctx context.Context, memgraphha *memgraphv1.MemgraphHA, logger *logr.Logger, dataInstanceId int) (bool, error) {
-	name := fmt.Sprintf("memgraph-data-%d", dataInstanceId)
-	logger.Info("Started reconciling", name)
-	dataInstanceStatefulSet := &appsv1.StatefulSet{}
-	err := r.Get(ctx, types.NamespacedName{Name: name, Namespace: memgraphha.Namespace}, dataInstanceStatefulSet)
-
-	if err == nil {
-		logger.Info("StatefulSet", name, "already exists.")
-		return false, nil
-	}
-
-	if errors.IsNotFound(err) {
-		dataInstance := r.createStatefulSetForDataInstance(memgraphha, dataInstanceId)
-		logger.Info("Creating a new StatefulSet", "StatefulSet.Namespace", dataInstance.Namespace, "StatefulSet.Name", dataInstance.Name)
-		err := r.Create(ctx, dataInstance)
-		if err != nil {
-			logger.Error(err, "Failed to create new StatefulSet", "StatefulSet.Namespace", dataInstance.Namespace, "StatefulSet.Name", dataInstance.Name)
-			return true, err
-		}
-		logger.Info("StatefulSet", name, "is created.")
-		return true, nil
-	}
-
-	logger.Error(err, "Failed to fetch StatefulSet", name)
-	return true, err
-
-}
-
-func (r *MemgraphHAReconciler) createStatefulSetForDataInstance(memgraphha *memgraphv1.MemgraphHA, dataInstanceId int) *appsv1.StatefulSet {
-	dataInstanceName := fmt.Sprintf("memgraph-data-%d", dataInstanceId)
-	labels := createDataInstanceLabels(dataInstanceName)
-	replicas := int32(1)
-	containerName := "memgraph-data"
-	image := "memgraph/memgraph:2.18.1"
-	boltPort := 7687
-	replicationPort := 12000
-	mgmtPort := 10000
-	args := []string{
-		fmt.Sprintf("--management-port=%d", mgmtPort),
-		fmt.Sprintf("--bolt-port=%d", boltPort),
-		"--experimental-enabled=high-availability",
-		"--also-log-to-stderr",
-		"--log-level=TRACE",
-		"--log-file=/var/log/memgraph/memgraph.log",
-	}
-	license := "<TODO> add"
-	organization := "testing-k8"
-	volumeLibName := fmt.Sprintf("%s-lib-storage", dataInstanceName)
-	volumeLibSize := "1Gi"
-	volumeLogName := fmt.Sprintf("%s-log-storage", dataInstanceName)
-	volumeLogSize := "256Mi"
-	initContainerName := "init"
-	initContainerCommand := []string{
-		"/bin/sh",
-		"-c",
-	}
-	initContainerArgs := []string{"chown -R memgraph:memgraph /var/log; chown -R memgraph:memgraph /var/lib"}
-	initContainerPrivileged := true
-	initContainerReadOnlyRootFilesystem := false
-	initContainerRunAsNonRoot := false
-	initContainerRunAsUser := int64(0)
-
-	// TODO:
-	/*
-		add serviceName
-		env
-	*/
-
-	data := &appsv1.StatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      dataInstanceName,
-			Namespace: memgraphha.Namespace,
-		},
-		Spec: appsv1.StatefulSetSpec{
-			Replicas: &replicas,
-			Selector: &metav1.LabelSelector{
-				MatchLabels: labels,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
-				},
-				Spec: corev1.PodSpec{
-					InitContainers: []corev1.Container{
-						{
-							Name:  initContainerName,
-							Image: image,
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									Name:      volumeLibName,
-									MountPath: "/var/lib/memgraph",
-								},
-								{
-									Name:      volumeLogName,
-									MountPath: "/var/log/memgraph",
-								},
-							},
-							Command: initContainerCommand,
-							Args:    initContainerArgs,
-							SecurityContext: &corev1.SecurityContext{
-								Privileged:             &initContainerPrivileged,
-								ReadOnlyRootFilesystem: &initContainerReadOnlyRootFilesystem,
-								RunAsNonRoot:           &initContainerRunAsNonRoot,
-								RunAsUser:              &initContainerRunAsUser,
-								Capabilities: &corev1.Capabilities{
-									Drop: []corev1.Capability{"all"},
-									Add:  []corev1.Capability{"CHOWN"},
-								},
-							},
-						},
-					},
-
-					Containers: []corev1.Container{{
-						Name:            containerName,
-						Image:           image,
-						ImagePullPolicy: corev1.PullIfNotPresent,
-						Ports: []corev1.ContainerPort{
-							{
-								ContainerPort: int32(boltPort),
-								Name:          "boltPort",
-							},
-							{
-								ContainerPort: int32(mgmtPort),
-								Name:          "managementPort",
-							},
-							{
-								ContainerPort: int32(replicationPort),
-								Name:          "replicationPort",
-							},
-						},
-						Args: args,
-						Env: []corev1.EnvVar{
-							{
-								Name:  "MEMGRAPH_ENTERPRISE_LICENSE",
-								Value: license,
-							},
-							{
-								Name:  "MEMGRAPH_ORGANIZATION_NAME",
-								Value: organization,
-							},
-						},
-						VolumeMounts: []corev1.VolumeMount{
-							{
-								Name:      volumeLibName,
-								MountPath: "/var/lib/memgraph",
-							},
-							{
-								Name:      volumeLogName,
-								MountPath: "/var/log/memgraph",
-							},
-						},
-					}},
-				},
-			},
-			VolumeClaimTemplates: []corev1.PersistentVolumeClaim{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: volumeLibName,
-					},
-					Spec: corev1.PersistentVolumeClaimSpec{
-						AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-						Resources: corev1.VolumeResourceRequirements{
-							Requests: corev1.ResourceList{
-								corev1.ResourceEphemeralStorage: resource.MustParse(volumeLibSize),
-							},
-						},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: volumeLogName,
-					},
-					Spec: corev1.PersistentVolumeClaimSpec{
-						AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-						Resources: corev1.VolumeResourceRequirements{
-							Requests: corev1.ResourceList{
-								corev1.ResourceEphemeralStorage: resource.MustParse(volumeLogSize),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	ctrl.SetControllerReference(memgraphha, data, r.Scheme)
-	return data
-}
-
-func (r *MemgraphHAReconciler) reconcileCoordNodePortService(ctx context.Context, memgraphha *memgraphv1.MemgraphHA, logger *logr.Logger, coordId int) (bool, error) {
-	serviceName := fmt.Sprintf("memgraph-coordinator-%d-external", coordId)
-	logger.Info("Started reconciling NodePort service", serviceName)
-
-	coordNodePortService := &corev1.Service{}
-	err := r.Get(ctx, types.NamespacedName{Name: serviceName, Namespace: memgraphha.Namespace}, coordNodePortService)
-
-	if err == nil {
-		logger.Info("NodePort", serviceName, "already exists.")
-		return false, nil
-	}
-
-	if errors.IsNotFound(err) {
-		nodePort := r.createCoordNodePort(memgraphha, coordId)
-		logger.Info("Creating a new NodePort", "NodePort.Namespace", nodePort.Namespace, "NodePort.Name", nodePort.Name)
-		err := r.Create(ctx, nodePort)
-		if err != nil {
-			logger.Error(err, "Failed to create new NodePort", "NodePort.Namespace", nodePort.Namespace, "NodePort.Name", nodePort.Name)
-			return true, err
-		}
-		logger.Info("NodePort", serviceName, "is created.")
-		return true, nil
-	}
-
-	logger.Error(err, "Failed to fetch NodePort", serviceName)
-	return true, err
-
-}
-
-func (r *MemgraphHAReconciler) createCoordNodePort(memgraphha *memgraphv1.MemgraphHA, coordId int) *corev1.Service {
-	serviceName := fmt.Sprintf("memgraph-coordinator-%d-external", coordId)
-	coordName := fmt.Sprintf("memgraph-coordinator-%d", coordId)
-	// TODO: (andi) Extract somehow configuration and move into separate files.
-	boltPort := 7687
-
-	coordNodePort := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serviceName,
-			Namespace: memgraphha.Namespace,
-		},
-		Spec: corev1.ServiceSpec{
-			Type:     corev1.ServiceTypeNodePort,
-			Selector: createCoordLabels(coordName),
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "bolt",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(boltPort),
-					TargetPort: intstr.FromInt(boltPort),
-				},
-			},
-		},
-	}
-
-	ctrl.SetControllerReference(memgraphha, coordNodePort, r.Scheme)
-	return coordNodePort
-}
-
-func (r *MemgraphHAReconciler) reconcileCoordClusterIPService(ctx context.Context, memgraphha *memgraphv1.MemgraphHA, logger *logr.Logger, coordId int) (bool, error) {
-	serviceName := fmt.Sprintf("memgraph-coordinator-%d", coordId)
-	logger.Info("Started reconciling ClusterIP service", serviceName)
-
-	coordClusterIPService := &corev1.Service{}
-	err := r.Get(ctx, types.NamespacedName{Name: serviceName, Namespace: memgraphha.Namespace}, coordClusterIPService)
-
-	if err == nil {
-		logger.Info("ClusterIP", serviceName, "already exists.")
-		return false, nil
-	}
-
-	if errors.IsNotFound(err) {
-		clusterIP := r.createCoordClusterIP(memgraphha, coordId)
-		logger.Info("Creating a new ClusterIP", "ClusterIP.Namespace", clusterIP.Namespace, "ClusterIP.Name", clusterIP.Name)
-		err := r.Create(ctx, clusterIP)
-		if err != nil {
-			logger.Error(err, "Failed to create new ClusterIP", "ClusterIP.Namespace", clusterIP.Namespace, "ClusterIP.Name", clusterIP.Name)
-			return true, err
-		}
-		logger.Info("ClusterIP", serviceName, "is created.")
-		return true, nil
-	}
-
-	logger.Error(err, "Failed to fetch ClusterIP", serviceName)
-	return true, err
-
-}
-
-func (r *MemgraphHAReconciler) createCoordClusterIP(memgraphha *memgraphv1.MemgraphHA, coordId int) *corev1.Service {
-	serviceName := fmt.Sprintf("memgraph-coordinator-%d", coordId)
-	coordName := serviceName
-	// TODO: (andi) Extract somehow configuration and move into separate files.
-	boltPort := 7687
-	coordinatorPort := 12000
-	mgmtPort := 10000
-
-	coordClusterIP := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serviceName,
-			Namespace: memgraphha.Namespace,
-		},
-		Spec: corev1.ServiceSpec{
-			Type:     corev1.ServiceTypeClusterIP,
-			Selector: createCoordLabels(coordName),
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "bolt",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(boltPort),
-					TargetPort: intstr.FromInt(boltPort),
-				},
-				{
-					Name:       "coordinator",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(coordinatorPort),
-					TargetPort: intstr.FromInt(coordinatorPort),
-				},
-				{
-					Name:       "management",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(mgmtPort),
-					TargetPort: intstr.FromInt(mgmtPort),
-				},
-			},
-		},
-	}
-
-	ctrl.SetControllerReference(memgraphha, coordClusterIP, r.Scheme)
-	return coordClusterIP
-}
-
-/*
-Returns bool, error tuple. If error exists, the caller should return with error and status will always be set to true.
-If there is no error, we must look at bool status which when true will say that the coordinator was createdand we need to requeue
-or that nothing was done and we can continue with the next step of reconciliation.
-*/
-func (r *MemgraphHAReconciler) reconcileCoordinator(ctx context.Context, memgraphha *memgraphv1.MemgraphHA, logger *logr.Logger, coordId int) (bool, error) {
-	name := fmt.Sprintf("memgraph-coordinator-%d", coordId)
-	logger.Info("Started reconciling", name)
-	coordStatefulSet := &appsv1.StatefulSet{}
-	err := r.Get(ctx, types.NamespacedName{Name: name, Namespace: memgraphha.Namespace}, coordStatefulSet)
-
-	if err == nil {
-		logger.Info("StatefulSet", name, "already exists.")
-		return false, nil
-	}
-
-	if errors.IsNotFound(err) {
-		coord := r.createStatefulSetForCoord(memgraphha, coordId)
-		logger.Info("Creating a new StatefulSet", "StatefulSet.Namespace", coord.Namespace, "StatefulSet.Name", coord.Name)
-		err := r.Create(ctx, coord)
-		if err != nil {
-			logger.Error(err, "Failed to create new StatefulSet", "StatefulSet.Namespace", coord.Namespace, "StatefulSet.Name", coord.Name)
-			return true, err
-		}
-		logger.Info("StatefulSet", name, "is created.")
-		return true, nil
-	}
-
-	logger.Error(err, "Failed to fetch StatefulSet", name)
-	return true, err
-}
-
-func (r *MemgraphHAReconciler) createStatefulSetForCoord(memgraphha *memgraphv1.MemgraphHA, coordId int) *appsv1.StatefulSet {
-	coordName := fmt.Sprintf("memgraph-coordinator-%d", coordId)
-	serviceName := coordName
-	labels := createCoordLabels(coordName)
-	replicas := int32(1)
-	containerName := "memgraph-coordinator"
-	image := "memgraph/memgraph:2.18.1"
-	boltPort := 7687
-	coordPort := 12000
-	mgmtPort := 10000
-	args := []string{
-		fmt.Sprintf("--coordinator-id=%d", coordId),
-		fmt.Sprintf("--coordinator-port=%d", coordPort),
-		fmt.Sprintf("--management-port=%d", mgmtPort),
-		fmt.Sprintf("--bolt-port=%d", boltPort),
-		fmt.Sprintf("--coordinator-hostname=%s.default.svc.cluster.local", coordName),
-		"--experimental-enabled=high-availability",
-		"--also-log-to-stderr",
-		"--log-level=TRACE",
-		"--log-file=/var/log/memgraph/memgraph.log",
-		"--nuraft-log-file=/var/log/memgraph/memgraph.log",
-	}
-	license := "<TODO> add"
-	organization := "testing-k8"
-	volumeLibName := fmt.Sprintf("%s-lib-storage", coordName)
-	volumeLibSize := "1Gi"
-	volumeLogName := fmt.Sprintf("%s-log-storage", coordName)
-	volumeLogSize := "256Mi"
-	initContainerName := "init"
-	initContainerCommand := []string{
-		"/bin/sh",
-		"-c",
-	}
-	initContainerArgs := []string{"chown -R memgraph:memgraph /var/log; chown -R memgraph:memgraph /var/lib"}
-	initContainerPrivileged := true
-	initContainerReadOnlyRootFilesystem := false
-	initContainerRunAsNonRoot := false
-	initContainerRunAsUser := int64(0)
-
-	// TODO:
-	/*
-		add serviceName
-		env
-	*/
-
-	coord := &appsv1.StatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      coordName,
-			Namespace: memgraphha.Namespace,
-		},
-		Spec: appsv1.StatefulSetSpec{
-			ServiceName: serviceName,
-			Replicas:    &replicas,
-			Selector: &metav1.LabelSelector{
-				MatchLabels: labels,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
-				},
-				Spec: corev1.PodSpec{
-					InitContainers: []corev1.Container{
-						{
-							Name:  initContainerName,
-							Image: image,
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									Name:      volumeLibName,
-									MountPath: "/var/lib/memgraph",
-								},
-								{
-									Name:      volumeLogName,
-									MountPath: "/var/log/memgraph",
-								},
-							},
-							Command: initContainerCommand,
-							Args:    initContainerArgs,
-							SecurityContext: &corev1.SecurityContext{
-								Privileged:             &initContainerPrivileged,
-								ReadOnlyRootFilesystem: &initContainerReadOnlyRootFilesystem,
-								RunAsNonRoot:           &initContainerRunAsNonRoot,
-								RunAsUser:              &initContainerRunAsUser,
-								Capabilities: &corev1.Capabilities{
-									Drop: []corev1.Capability{"all"},
-									Add:  []corev1.Capability{"CHOWN"},
-								},
-							},
-						},
-					},
-
-					Containers: []corev1.Container{{
-						Name:            containerName,
-						Image:           image,
-						ImagePullPolicy: corev1.PullIfNotPresent,
-						Ports: []corev1.ContainerPort{
-							{
-								ContainerPort: int32(boltPort),
-								Name:          "boltPort",
-							},
-							{
-								ContainerPort: int32(mgmtPort),
-								Name:          "managementPort",
-							},
-							{
-								ContainerPort: int32(coordPort),
-								Name:          "coordinatorPort",
-							},
-						},
-						Args: args,
-						Env: []corev1.EnvVar{
-							{
-								Name:  "MEMGRAPH_ENTERPRISE_LICENSE",
-								Value: license,
-							},
-							{
-								Name:  "MEMGRAPH_ORGANIZATION_NAME",
-								Value: organization,
-							},
-						},
-						VolumeMounts: []corev1.VolumeMount{
-							{
-								Name:      volumeLibName,
-								MountPath: "/var/lib/memgraph",
-							},
-							{
-								Name:      volumeLogName,
-								MountPath: "/var/log/memgraph",
-							},
-						},
-					}},
-				},
-			},
-			VolumeClaimTemplates: []corev1.PersistentVolumeClaim{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: volumeLibName,
-					},
-					Spec: corev1.PersistentVolumeClaimSpec{
-						AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-						Resources: corev1.VolumeResourceRequirements{
-							Requests: corev1.ResourceList{
-								corev1.ResourceEphemeralStorage: resource.MustParse(volumeLibSize),
-							},
-						},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: volumeLogName,
-					},
-					Spec: corev1.PersistentVolumeClaimSpec{
-						AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-						Resources: corev1.VolumeResourceRequirements{
-							Requests: corev1.ResourceList{
-								corev1.ResourceEphemeralStorage: resource.MustParse(volumeLogSize),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	ctrl.SetControllerReference(memgraphha, coord, r.Scheme)
-	return coord
-}
-
-func createCoordLabels(coordName string) map[string]string {
-	return map[string]string{"app": coordName}
-}
-
-func createDataInstanceLabels(dataInstanceName string) map[string]string {
-	return map[string]string{"app": dataInstanceName}
 }
 
 // SetupWithManager sets up the controller with the Manager.
